Added Stream API realization

+++ src/main/java/repository/WriterRepository.java
@@ -2,6 +2,6 @@ package repository;
 
 import model.Writer;
 
-public interface WriterRepository extends GenericRepository<Writer, Long>{
+public interface WriterRepository extends GenericRepository<Writer, Integer>{
 
 }
